+++ APICurso.Domain/Entities/LimiteCliente.cs
@@ -4,7 +4,7 @@
     {
         //Os private set serve para a proteger de alteraçoes o cliente sem que seja gerado log
         public int Codigo { get; private set; } 
-        public double LimiteCredito { get; private set; }
+        public decimal LimiteCredito { get; private set; }
         public string Nome { get; private set; }
         public bool IsActive { get; private set; }
 
@@ -14,7 +14,7 @@
         }
 
         //metodo para Alterar o limite de credito
-        public void AlterarLimiteDeCredito(double valor, bool subtrair)
+        public void AlterarLimiteDeCredito(decimal valor, bool subtrair)
         {
             if (subtrair)//Verifica se é para subtrair ou somar ao limite de credito já existente
             {
@@ -27,7 +27,7 @@
         }
 
         //metodo para cadastrar novo cliente
-        public void CadastrarNovoCliente(string nome, double limite, int codigo)
+        public void CadastrarNovoCliente(string nome, decimal limite, int codigo)
         {
             this.Codigo = codigo;
             this.Nome = nome;
